[Update] AboutCounting 컴포넌트화
