+++ src/Components/About/index.tsx
@@ -1,47 +1,48 @@
-import { PageExplanation } from "../PageExplanation";
+// import { PageExplanation } from "../PageExplanation";
+import AboutCounting from "Components/AboutCounting";
 import * as S from "./style";
-import Table from "Api/table";
-import { useEffect, useState } from "react";
+// import Table from "Api/table";
+// import { useEffect, useState } from "react";
 
-const useAmount = (): [amount: number, date: number] => {
-  const [amount, setAmount] = useState({ data: 0 });
-  const [date, setDate] = useState({ data: 0 });
-  const getAmount = async () => {
-    return await Table.amountUncomfortable();
-  };
+// const useAmount = (): [amount: number, date: number] => {
+//   const [amount, setAmount] = useState({ data: 0 });
+//   const [date, setDate] = useState({ data: 0 });
+//   const getAmount = async () => {
+//     return await Table.amountUncomfortable();
+//   };
 
-  const getDataSinceProjectStart = async () => {
-    return await Table.dateSinceProjectStart();
-  };
+//   const getDataSinceProjectStart = async () => {
+//     return await Table.dateSinceProjectStart();
+//   };
 
-  useEffect(() => {
-    getAmount().then((res) => setAmount(res.data));
-    getDataSinceProjectStart().then((res) => setDate(res.data));
-  }, []);
+//   useEffect(() => {
+//     getAmount().then((res) => setAmount(res.data));
+//     getDataSinceProjectStart().then((res) => setDate(res.data));
+//   }, []);
 
-  return [amount.data, date.data];
-};
+//   return [amount.data, date.data];
+// };
 
-const useHeading = (): string[][] => {
-  const [amount, date] = useAmount();
-  const day: number = date;
-  const cnt_comment: number = amount;
+// const useHeading = (): string[][] => {
+//   const [amount, date] = useAmount();
+//   const day: number = date;
+//   const cnt_comment: number = amount;
 
-  const heading: string[] = [
-    `${day}일 동안`,
-    `${cnt_comment}개의 불편함이`,
-    `모였습니다.`,
-  ];
-  const explanation: string[] = [""];
+//   const heading: string[] = [
+//     `${day}일 동안`,
+//     `${cnt_comment}개의 불편함이`,
+//     `모였습니다.`,
+//   ];
+//   const explanation: string[] = [""];
 
-  return [heading, explanation];
-};
+//   return [heading, explanation];
+// };
 
 const AboutPage: React.FC = () => {
-  const [heading, explanation] = useHeading();
+  // const [heading, explanation] = useHeading();
   return (
     <S.AboutBox>
-      <PageExplanation heading={heading} explanation={explanation} />
+      <AboutCounting></AboutCounting>
       <S.AboutHeading>학교가 불편한 순간</S.AboutHeading>
       <S.AboutContent>
         <p>
