+++ src/Components/AboutCounting/style.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const Counting = styled.div`
+  font-weight: 700;
+  font-size: 30px;
+  color: #6a76e9;
+  margin-bottom: 50px;
+  p{
+    margin: 5px 0;
+  }
+`;
